% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiences-package.R
\docType{package}
\name{experiences-package}
\alias{experiences}
\alias{experiences-package}
\title{experiences: Experience Research}
\description{
Provides convenience functions for researching experiences. The functions are designed to translate statistical approaches to applied experience research.
}
\author{
\strong{Maintainer}: Joe Chelladurai \email{joe.chelladurai@outlook.com} (\href{https://orcid.org/0000-0001-8477-3753}{ORCID})

}
\keyword{internal}
