% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_dist_two_tailed.R
\name{t_dist_two_tailed}
\alias{t_dist_two_tailed}
\title{T distribution - two-tailed}
\usage{
t_dist_two_tailed(t_score, degrees_of_freedom)
}
\arguments{
\item{t_score}{t value}

\item{degrees_of_freedom}{degrees of freedom}
}
\value{
value
}
\description{
T distribution - two-tailed
}
